requirements.txt
pytest
pytest-env
pytest-order
pytest-playwright